# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyTailz API",
    "version": "1.0.0",
    "description": "Local Swagger documentation for HappyTailz backend"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://happytailz-9qf9.onrender.com/", "description": "Render" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/api/userplans": {
      "get": {
        "summary": "List my user plans",
        "tags": ["UserPlans"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Subscribe to a plan",
        "tags": ["UserPlans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["planId"],
                "properties": {
                  "planId": { "type": "string" },
                  "startDate": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/walkingrequests": {
      "post": {
        "summary": "Create walking request",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pet", "scheduleAt"],
                "properties": {
                  "pet": { "type": "string" },
                  "plan": { "type": "string" },
                  "scheduleAt": { "type": "string", "format": "date-time" },
                  "location": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/walkingrequests/{id}/approve": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "Approve a walking request (admin)",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Approved" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/walkingrequests/{id}/assign": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "Assign a driver to a walking request (admin)",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["driver"],
                "properties": {
                  "driver": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Assigned" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/walkingrequests/{id}/complete": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "Mark walking request as completed (driver)",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Completed" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/walkingrequests/me": {
      "get": {
        "summary": "List my walking requests",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/profile": {
      "get": {
        "summary": "Get my profile",
        "tags": ["Profile"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "put": {
        "summary": "Update my profile",
        "tags": ["Profile"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      }
    },
    "/api/profile/password": {
      "put": {
        "summary": "Change my password",
        "tags": ["Profile"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["oldPassword", "newPassword"],
                "properties": {
                  "oldPassword": { "type": "string" },
                  "newPassword": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Bad Request" }, "404": { "description": "Not Found" } }
      }
    },
...
    "/api/admin/plans": {
      "get": { "summary": "List plans", "tags": ["Admin - Plans"], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create plan",
        "tags": ["Admin - Plans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "durationDays", "price"],
                "properties": {
                  "name": { "type": "string", "enum": ["1 day", "1 week", "1 month"] },
                  "durationDays": { "type": "integer" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/plans/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update plan",
        "tags": ["Admin - Plans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "enum": ["1 day", "1 week", "1 month"] },
                  "durationDays": { "type": "integer" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete plan", "tags": ["Admin - Plans"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List all users",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "driver", "user"] } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search by name/email" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "email": { "type": "string" },
                      "role": { "type": "string" },
                      "activePlan": { "type": "object", "nullable": true }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "409": { "description": "Email exists" } }
      }
    },
    "/api/admin/users/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "summary": "Get user by id",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "role": { "type": "string" },
                    "activePlan": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Update user",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user"] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user"] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete user", "tags": ["Admin - Users"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
...
  }
}